FAC-993: Refactor - Move SpaceStatus and SpaceStatusTest to common module
