+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/domain/Space.java
@@ -2,6 +2,7 @@ package org.flickit.assessment.core.application.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.flickit.assessment.common.application.domain.space.SpaceStatus;
 
 @Getter
 @RequiredArgsConstructor
@@ -9,4 +10,5 @@ public class Space {
 
     private final long id;
     private final String title;
+    private final SpaceStatus status;
 }
